news_scraper/backend/requirements.txt
# API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.9.2
pydantic-settings==2.1.0

# Database
pymongo==4.6.1
motor==3.3.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# AI/Video Generation
google-generativeai==0.3.0

# Scheduler
schedule==1.2.1
